Creado el trait AuxiliarFunctions para generalizar el uso de funciones auxiliares genéricas y evitar duplicidad de código. Se implementa el método getProfilePhotos() que se encontraba en el componente de Livewire UpdateProfileInformationForm. Se implementa el trait en dicho componente y en UpdateUserProfileInfirmation dónde se implementará la funcionalidad de sólo permitir un máximo de 5 fotos de perfil y si sube una sexta se elimina la foto de perfil más antigua.
